trunk/requirements.txt
PyYAML==3.12
requests==2.13.0
deepdiff==3.1.0
